Cleanup: Move memory-related sources to their own folder.
Created a new folder ("src/memory/") to group memory-related
files (memory.c, mpu.c); we can add another source for SmartEEPROM
functionality in here.
