+++ src/backup.c
@@ -19,7 +19,7 @@
 #include <backup.h>
 #include <hardfault.h>
 #include <keystore.h>
-#include <memory.h>
+#include <memory/memory.h>
 #include <pb_decode.h>
 #include <pb_encode.h>
 #include <restore.h>
