+++ src/commander/commander.c
@@ -28,7 +28,7 @@
 
 #include <flags.h>
 #include <hardfault.h>
-#include <memory.h>
+#include <memory/memory.h>
 #include <random.h>
 #include <screen.h>
 #include <sd.h>
