+++ src/common_main.c
@@ -17,8 +17,8 @@
 #include "flags.h"
 #include "hardfault.h"
 #include "keystore.h"
-#include "memory.h"
-#include "mpu.h"
+#include "memory/memory.h"
+#include "memory/mpu.h"
 #include "random.h"
 #include "screen.h"
 #include "securechip/securechip.h"
