+++ src/workflow/backup.c
@@ -15,7 +15,7 @@
 #include "workflow.h"
 #include <backup.h>
 #include <hww.pb.h>
-#include <memory.h>
+#include <memory/memory.h>
 #include <sd.h>
 #include <stdio.h>
 #include <time.h>
