+++ src/workflow/unlock.c
@@ -18,7 +18,7 @@
 #include "workflow.h"
 #include <hardfault.h>
 #include <keystore.h>
-#include <memory.h>
+#include <memory/memory.h>
 #include <screen.h>
 #include <string.h>
 #include <ui/components/ui_images.h>
