+++ test/device-test/src/test_bip39.c
@@ -26,7 +26,6 @@
 
 #include "hardfault.h"
 #include "keystore.h"
-#include "memory.h"
 #include "random.h"
 #include "screen.h"
 #include "sd.h"
