+++ test/device-test/src/test_usb_composed_ep_out.c
@@ -14,7 +14,6 @@
 
 #include "common_main.h"
 #include "driver_init.h"
-#include "memory.h"
 #include "qtouch.h"
 #include "random.h"
 #include "screen.h"
