+++ test/device-test/src/test_usbdc_init.c
@@ -14,7 +14,6 @@
 
 #include "common_main.h"
 #include "driver_init.h"
-#include "memory.h"
 #include "qtouch.h"
 #include "random.h"
 #include "screen.h"
