+++ test/unit-test/test_u2f_hid.c
@@ -17,9 +17,9 @@
 #include <sys/ioctl.h>
 #endif
 
-#include "memory.h"
 #include "random.h"
 #include "u2f/ecc.h"
+#include "util.h"
 
 #include "u2f.h"
 #include "u2f/u2f_util_t.h"
